% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretmodeleval.R
\name{modeleval}
\alias{modeleval}
\title{Caret model evaulation}
\usage{
modeleval(target, data, timeouttime = 60, metric, holdout = 0.8,
  trainmethod = "repeatedcv", number = 5, repeats = 3,
  order = "accending")
}
\arguments{
\item{target}{the target vaiable is characture format}

\item{data}{the data frame of intrest}

\item{timeouttime}{the time in seconds each mode is permitted to run}

\item{metric}{metric of accuracy possibe terms are RMSE, R2, MAE for regression, and Accuracy or Kappa for classification}

\item{holdout}{hold out percentage for the training set, .8 is 80% of data used for training}

\item{number}{in train control how many number}

\item{repeats}{in train control how many repeats}

\item{order}{the order of sorting}

\item{trainmetod}{method of training ex repeatedcv}
}
\value{
a dataframe with model and accuracy metric sorted
}
\description{
Caret model evaulation
}
\examples{




}
